spark agent #33 small fixes, refactoring
