+++ core/src/main/scala/za/co/absa/spline/harvester/builder/OperationNodeBuilder.scala
@@ -16,12 +16,9 @@
 
 package za.co.absa.spline.harvester.builder
 
-import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.catalyst.{expressions => sparkExprssions}
-import za.co.absa.commons.lang.CachingConverter
 import za.co.absa.spline.harvester.IdGeneratorsBundle
-import za.co.absa.spline.harvester.builder.PlanOperationNodeBuilder.OperationId
-import za.co.absa.spline.harvester.converter._
+import za.co.absa.spline.harvester.builder.plan.PlanOperationNodeBuilder.OperationId
 import za.co.absa.spline.producer.model.{Attribute, FunctionalExpression, Literal}
 
 trait OperationNodeBuilder {
