+++ integration-tests/data/cities.csv
@@ -1,4 +1,4 @@
 name,population
 Prague,1275406
-ape Town,4710000
+Cape Town,4710000
 Johannesburg,5635127
